+++ src/main/kotlin/com/github/zero9178/mlirods/language/stubs/impl/TableGenIdentifierElementStub.kt
@@ -2,6 +2,8 @@ package com.github.zero9178.mlirods.language.stubs.impl
 
 import com.github.zero9178.mlirods.index.ALL_IDENTIFIERS_INDEX
 import com.github.zero9178.mlirods.index.IDENTIFIER_INDEX
+import com.github.zero9178.mlirods.index.MAY_DERIVE_CLASS_INDEX
+import com.github.zero9178.mlirods.language.generated.psi.TableGenDefStatement
 import com.github.zero9178.mlirods.language.generated.psi.impl.TableGenDefStatementImpl
 import com.github.zero9178.mlirods.language.generated.psi.impl.TableGenDefvarStatementImpl
 import com.github.zero9178.mlirods.language.psi.TableGenIdentifierElement
@@ -13,7 +15,7 @@ import com.intellij.psi.stubs.*
 /**
  * Stub interface for [TableGenIdentifierElement] elements.
  */
-interface TableGenIdentifierElementStub : StubElement<TableGenIdentifierElement> {
+sealed interface TableGenIdentifierElementStub : StubElement<TableGenIdentifierElement> {
     val name: String?
 }
 
@@ -45,6 +47,14 @@ abstract class TableGenAbstractIdentifierElementStubElementType(
     }
 }
 
+private class TableGenIdentifierElementStubImpl(
+    override val name: String?,
+    parent: StubElement<out PsiElement>?,
+    elementType: TableGenAbstractIdentifierElementStubElementType,
+) : StubBase<TableGenIdentifierElement>(
+    parent, elementType
+), TableGenIdentifierElementStub
+
 class TableGenDefvarStatementStubElementType(debugName: String) : TableGenAbstractIdentifierElementStubElementType(
     debugName,
     ::TableGenDefvarStatementImpl
@@ -65,25 +75,63 @@ class TableGenDefvarStatementStubElementType(debugName: String) : TableGenAbstra
     }
 }
 
-class TableGenDefStatementStubElementType(debugName: String) : TableGenAbstractIdentifierElementStubElementType(
+sealed interface TableGenDefStatementStub : TableGenIdentifierElementStub {
+    val baseClassNames: List<String>
+}
+
+class TableGenDefStatementStubElementType(debugName: String) :
+    TableGenStubElementType<TableGenDefStatementStub, TableGenDefStatement>(
     debugName,
     ::TableGenDefStatementImpl
 ) {
+    override fun createStub(
+        psi: TableGenDefStatement, parentStub: StubElement<out PsiElement?>?
+    ): TableGenDefStatementStub {
+        return TableGenDefStatementElementStubImpl(psi.name, psi.classRefList.map { it.className }, parentStub, this)
+    }
+
+    override fun serialize(
+        stub: TableGenDefStatementStub, dataStream: StubOutputStream
+    ) {
+        dataStream.writeName(stub.name)
+        dataStream.writeVarInt(stub.baseClassNames.size)
+        stub.baseClassNames.forEach {
+            dataStream.writeName(it)
+        }
+    }
+
+    override fun deserialize(
+        dataStream: StubInputStream, parentStub: StubElement<*>?
+    ): TableGenDefStatementStub {
+        return TableGenDefStatementElementStubImpl(
+            dataStream.readNameString(),
+            buildList {
+                repeat(dataStream.readVarInt()) {
+                    add(dataStream.readNameString() ?: return@repeat)
+                }
+            }, parentStub, this
+        )
+    }
+
     override fun indexStub(
-        stub: TableGenIdentifierElementStub,
+        stub: TableGenDefStatementStub,
         sink: IndexSink
     ) {
         stub.name?.let {
             sink.occurrence(IDENTIFIER_INDEX, it)
         }
         sink.occurrence(ALL_IDENTIFIERS_INDEX, 0)
+        stub.baseClassNames.forEach {
+            sink.occurrence(MAY_DERIVE_CLASS_INDEX, it)
+        }
     }
 }
 
-private class TableGenIdentifierElementStubImpl(
+private class TableGenDefStatementElementStubImpl(
     override val name: String?,
+    override val baseClassNames: List<String>,
     parent: StubElement<out PsiElement>?,
-    elementType: TableGenAbstractIdentifierElementStubElementType,
+    elementType: TableGenDefStatementStubElementType,
 ) : StubBase<TableGenIdentifierElement>(
     parent, elementType
-), TableGenIdentifierElementStub
+), TableGenDefStatementStub